refactor(menu): separate menu code
 - add settings menu
