--- lib/app/view/app.dart
@@ -1,99 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:pomodoro_timer/l10n/l10n.dart';
-import 'package:pomodoro_timer/timer/view/timer_page.dart';
-import 'package:pomodoro_timer/tray/cubit/tray_cubit.dart';
-import 'package:window_manager/window_manager.dart';
-
-class App extends StatelessWidget {
-  const App({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      theme: ThemeData(
-        appBarTheme: const AppBarTheme(color: Color(0xFFFFFFFF)),
-        colorScheme: ColorScheme.fromSwatch(
-          accentColor: const Color(0xFFEE5436),
-        ),
-        useMaterial3: true,
-      ),
-      darkTheme: ThemeData.dark(useMaterial3: true),
-      themeMode: ThemeMode.dark,
-      localizationsDelegates: AppLocalizations.localizationsDelegates,
-      supportedLocales: AppLocalizations.supportedLocales,
-      home: BlocProvider(
-        create: (context) => TrayCubit()..init(),
-        child: PlatformMenuBar(
-          menus: [
-            PlatformMenu(
-              label: 'pomo_timer',
-              menus: [
-                PlatformMenuItemGroup(
-                  members: [
-                    if (PlatformProvidedMenuItem.hasMenu(
-                      PlatformProvidedMenuItemType.about,
-                    ))
-                      const PlatformProvidedMenuItem(
-                        type: PlatformProvidedMenuItemType.about,
-                      ),
-                  ],
-                ),
-                PlatformMenuItemGroup(
-                  members: [
-                    if (PlatformProvidedMenuItem.hasMenu(
-                      PlatformProvidedMenuItemType.servicesSubmenu,
-                    ))
-                      const PlatformProvidedMenuItem(
-                        type: PlatformProvidedMenuItemType.servicesSubmenu,
-                      ),
-                  ],
-                ),
-                PlatformMenuItemGroup(
-                  members: [
-                    if (PlatformProvidedMenuItem.hasMenu(
-                      PlatformProvidedMenuItemType.quit,
-                    ))
-                      const PlatformProvidedMenuItem(
-                        type: PlatformProvidedMenuItemType.quit,
-                      ),
-                  ],
-                ),
-              ],
-            ),
-            PlatformMenu(
-              label: 'View',
-              menus: [
-                if (PlatformProvidedMenuItem.hasMenu(
-                  PlatformProvidedMenuItemType.toggleFullScreen,
-                ))
-                  const PlatformProvidedMenuItem(
-                    type: PlatformProvidedMenuItemType.toggleFullScreen,
-                  ),
-              ],
-            ),
-            PlatformMenu(
-              label: 'Window',
-              menus: [
-                PlatformMenuItemGroup(
-                  members: [
-                    PlatformMenuItem(
-                      label: 'Close',
-                      shortcut: const SingleActivator(
-                        LogicalKeyboardKey.keyW,
-                        meta: true,
-                      ),
-                      onSelected: windowManager.close,
-                    ),
-                  ],
-                ),
-              ],
-            ),
-          ],
-          child: TimerPage(),
-        ),
-      ),
-    );
-  }
-}
