+++ lib/main.dart
@@ -2,5 +2,5 @@ import 'package:pomodoro_timer/app/app.dart';
 import 'package:pomodoro_timer/bootstrap.dart';
 
 void main() {
-  bootstrap(() => const App());
+  bootstrap(() => const AppPage());
 }
