+++ lib/services/material_context.dart
@@ -1,0 +1,3 @@
+import 'package:flutter/widgets.dart';
+
+final GlobalKey<NavigatorState> navigatorState = GlobalKey<NavigatorState>();
