+++ pubspec.yaml
@@ -15,6 +15,7 @@ dependencies:
   flutter_bloc: ^8.1.2
   flutter_localizations:
     sdk: flutter
+  hydrated_bloc: ^9.1.0
   intl: ^0.17.0
   just_audio: ^0.9.32
   launch_at_startup: ^0.2.2
