+++ test/app/view/app_test.dart
@@ -5,7 +5,7 @@ import 'package:pomodoro_timer/timer/view/timer_page.dart';
 void main() {
   group('App', () {
     testWidgets('renders CounterPage', (tester) async {
-      await tester.pumpWidget(const App());
+      await tester.pumpWidget(const AppPage());
       expect(find.byType(TimerPage), findsOneWidget);
     });
   });
