+++ test/widget_test.dart
@@ -7,12 +7,12 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:pomodoro_timer/app/view/app.dart';
+import 'package:pomodoro_timer/app/view/app_page.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const App());
+    await tester.pumpWidget(const AppPage());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
